Reorganize modules and imports
