+++ robotpose/__init__.py
@@ -1,0 +1,2 @@
+from .wizards import DatasetWizard
+from .data import Dataset, DatasetInfo
