+++ robotpose/data/__init__.py
@@ -1,0 +1,1 @@
+from .dataset import Dataset, DatasetInfo
