+++ view_pc.py
@@ -1,4 +1,4 @@
-from robotpose.dataset import Dataset
+from robotpose import Dataset
 import numpy as np
 import cv2
 import open3d as o3d
