+++ wizard.py
@@ -5,8 +5,7 @@ import sys
 os.environ['TF_CPP_MIN_LOG_LEVEL'] = '1'  # or any {'0', '1', '2'}
 import tensorflow as tf
 
-from robotpose.dataset import Dataset, DatasetInfo
-from robotpose.wizards import DatasetWizard
+from robotpose import DatasetWizard, Dataset, DatasetInfo
 
 
 if __name__ == "__main__":
